update:　Modify pngle to improve PNG decoding speed

+++ src/lgfx/v0/LGFXBase.cpp
@@ -2450,14 +2450,14 @@ namespace lgfx
     if (y1 > p->maxHeight) y1 = p->maxHeight;
     if (y0 >= y1) return;
 
-    while (x < p->offX && --len)
+    while ((int32_t)x < p->offX && --len)
     {
       x += div_x;
       rgba += 4;
     }
     x -= p->offX;
 
-    if (!len || x >= p->maxWidth) return;
+    if (!len || (int32_t)x >= p->maxWidth) return;
 
     size_t idx = 0;
     while ((rgba[idx * 4 + 3] == 255) && ++idx != len);
@@ -2481,7 +2481,7 @@ namespace lgfx
           }
         }
         x += div_x;
-        if (x >= p->maxWidth) break;
+        if ((int32_t)x >= p->maxWidth) break;
         rgba += 4;
       } while (--len);
       p->gfx->pushImage(p->x, p->y + y0, p->maxWidth, 1, p->pc, true);
@@ -2493,7 +2493,7 @@ namespace lgfx
       do
       {
         p->lineBuffer[x] = *(uint32_t*)rgba;
-        if (++x >= p->maxWidth) break;
+        if ((int32_t)++x >= p->maxWidth) break;
         rgba += 4;
       } while (--len);
       p->gfx->pushImage(p->x, p->y + y0, p->maxWidth, 1, p->pc, true);
@@ -2505,7 +2505,7 @@ namespace lgfx
         p->gfx->setColor(color888(rgba[0], rgba[1], rgba[2]));
         p->gfx->writeFillRectPreclipped(p->x + x, p->y + y0, 1, 1);
         x += div_x;
-        if (x >= p->maxWidth) break;
+        if ((int32_t)x >= p->maxWidth) break;
         rgba += 4;
       } while (--len);
     }
